+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Requires PHP `8.4`
 - Requires `innmind/filesystem:~9.0`
+- Requires `innmind/time:~1.0`
 
 ## 2.0.0 - 2025-04-16